+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-# 1.0.2 - 1.21.1
-- Added automated release tasks and changelog.
+# 1.1.0 - 1.21.1
+- Added command to open a chest menu for another player.
+    - `/ccm open <menu_id> [playername]`
+- Added `craft_items` action to craft outputs, based on inputs.